Changed the tenant_id in migration to bigint

+++ db/migrate/20151435234623_add_tenant_id_to_miq_ae_namespaces.rb
@@ -1,5 +1,5 @@
 class AddTenantIdToMiqAeNamespaces < ActiveRecord::Migration
   def change
-    add_column :miq_ae_namespaces, :tenant_id, :integer
+    add_column :miq_ae_namespaces, :tenant_id, :bigint
   end
 end
